requirements.txt
streamlit
chromadb
sentence-transformers
openai
python-dotenv
reportlab